+++ CHANGELOG.md
@@ -16,10 +16,12 @@ and this project adheres to
 - Add `is_trashed` flag to thread model
 - Add to select multiple threads in thread panel
 - Add image proxy endpoint to display external images in messages
+- Add `to_exact` modifier to search query
 
 ### Changed
 
 - Configure Drive App Name through environment variable (DRIVE_APP_NAME)
 - Inherit OIDC Authentication backend from django-lasuite #408
 - Exclude `is_trashed` and `is_spam` threads from search results by default
+- `to` search modifier now looks for messages where recipient fields (to, cc, bcc) contain the given email address.